requirements.txt
h5py
wget
torch
numpy
Pillow
matplotlib
torchvision
python-dateutil
opencv-contrib-python
